[CDF-22564] 🍖 Cleanup deploy/clean commands (#959)
* refactor: moved into core app

* docs: docstrings

* tests: updated tests part one

* tests: updated tests part two

* build: changelog

* refactor: added changes

* fix:

* fix: update

* Update CHANGELOG.cdf-tk.md
