+++ CHANGELOG.cdf-tk.md
@@ -41,6 +41,13 @@ Changes are grouped as follows:
   typically an object/dict, which caused loss of information. This is because
   `yaml.safe_dump(yaml.safe_load(config)) != config` as, for example, special YAML tags are lost.
 
+### Removed
+
+- The `--interactive` flag from the `cdf deploy` and `cdf clean` commands.
+- The shared flags `--verbose`. This has been deprecated and has been replaced by `--verbose` on each individual
+  command. For example, before you could write `cdf --verbose build --env dev`, now you should write
+  `cdf build --env dev --verbose`.
+
 ## [0.3.0a4] - 2024-09-13
 
 ### Added
