+++ cognite_toolkit/_cdf_tk/apps/__init__.py
@@ -1,7 +1,8 @@
 from ._auth_app import AuthApp
+from ._core_app import CoreApp
 from ._landing_app import LandingApp
 from ._modules_app import ModulesApp
 from ._repo_app import RepoApp
 from ._run import RunApp
 
-__all__ = ["LandingApp", "ModulesApp", "RunApp", "RepoApp", "AuthApp"]
+__all__ = ["CoreApp", "AuthApp", "LandingApp", "ModulesApp", "RunApp", "RepoApp"]
